requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
geoalchemy2==0.18.0
geopy==2.4.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
jinja2==3.1.6
python-multipart==0.0.20
pandas==2.3.2
python-dotenv==1.1.1
starlette==0.47.3
openpyxl==3.1.5
xlrd==2.0.1
email-validator==2.2.0
dash==2.17.1
